requirements.txt
ansible==2.7.10
asn1crypto==0.24.0
bcrypt==3.1.6
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
crypto==1.4.1
cryptography==2.6.1
idna==2.8
Jinja2==2.10.1
MarkupSafe==1.1.1
Naked==0.1.31
paramiko==2.4.2
pyasn1==0.4.5
pycparser==2.19
PyNaCl==1.3.0
PyYAML==5.1
requests==2.21.0
shellescape==3.4.1
six==1.12.0
urllib3==1.24.2
